+++ packages/AdminConfig/tests/Support/TestCase.php
@@ -31,6 +31,7 @@ abstract class TestCase extends BaseTestCase {
 		$GLOBALS['lerm_admin_config_is_multisite']      = false;
 		$GLOBALS['lerm_admin_config_options']           = array();
 		$GLOBALS['lerm_admin_config_rest_routes']       = array();
+		$GLOBALS['lerm_admin_config_current_user_can']  = true;
 		$GLOBALS['lerm_admin_config_ajax_nonce_checks'] = array();
 		$GLOBALS['lerm_admin_config_json_response']     = null;